requirements.txt
pycrypto
configparser
click
